Defensive cleanup of e107_make.php
